models: split provenance to submodules
