+++ renku/models/provenance/expanded.py
@@ -1,0 +1,38 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright 2018 - Swiss Data Science Center (SDSC)
+# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+# Eidgenössische Technische Hochschule Zürich (ETHZ).
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Provide additional terms that can be used to relate classes."""
+
+from renku.models import _jsonld as jsonld
+
+
+@jsonld.s(
+    type=[
+        'prov:Location',
+        'foaf:Project',
+    ],
+    context={
+        'foaf': 'http://xmlns.com/foaf/0.1/',
+        'prov': 'http://purl.org/dc/terms/',
+    },
+    frozen=True,
+    slots=True,
+)
+class Project(object):
+    """Represent a project."""
+
+    _id = jsonld.ib(context='@id', kw_only=True)
